# Fix performance-noexcept-swap clang-tidy warning