--- a/include/util/deallocating_vector.hpp
+++ b/include/util/deallocating_vector.hpp
@@ -166,7 +166,7 @@ class DeallocatingVectorIterator
 
 template <typename ElementT> class DeallocatingVector;
 
-template <typename T> void swap(DeallocatingVector<T> &lhs, DeallocatingVector<T> &rhs);
+template <typename T> void swap(DeallocatingVector<T> &lhs, DeallocatingVector<T> &rhs) noexcept;
 
 template <typename ElementT> class DeallocatingVector
 {
@@ -221,9 +221,10 @@ template <typename ElementT> class DeallocatingVector
 
     ~DeallocatingVector() { clear(); }
 
-    friend void swap<>(DeallocatingVector<ElementT> &lhs, DeallocatingVector<ElementT> &rhs);
+    friend void swap<>(DeallocatingVector<ElementT> &lhs,
+                       DeallocatingVector<ElementT> &rhs) noexcept;
 
-    void swap(DeallocatingVector<ElementT> &other)
+    void swap(DeallocatingVector<ElementT> &other) noexcept
     {
         std::swap(current_size, other.current_size);
         bucket_list.swap(other.bucket_list);
@@ -342,7 +343,7 @@ template <typename ElementT> class DeallocatingVector
     }
 };
 
-template <typename T> void swap(DeallocatingVector<T> &lhs, DeallocatingVector<T> &rhs)
+template <typename T> void swap(DeallocatingVector<T> &lhs, DeallocatingVector<T> &rhs) noexcept
 {
     lhs.swap(rhs);
 }