--- a/include/util/permutation.hpp
+++ b/include/util/permutation.hpp
@@ -10,9 +10,9 @@ namespace osrm::util
 
 namespace permutation_detail
 {
-template <typename T> static inline void swap(T &a, T &b) { std::swap(a, b); }
+template <typename T> static inline void swap(T &a, T &b) noexcept { std::swap(a, b); }
 
-static inline void swap(std::vector<bool>::reference a, std::vector<bool>::reference b)
+static inline void swap(std::vector<bool>::reference a, std::vector<bool>::reference b) noexcept
 {
     std::vector<bool>::swap(a, b);
 }