https://github.com/Project-OSRM/osrm-backend/issues/6928


<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1180.77<br/>plain u32: 1186.22<br/>aliased double: 1264.15<br/>plain double: 1244.71 | aliased u32: 1100.22<br/>plain u32: 1100.48<br/>aliased double: 958.534<br/>plain double: 952.62 |
| e2e_match_ch | requests: 362<br/>failures: 0<br/>req/s: 6.034req/s<br/>avg: 5.195ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 16ms<br/>98%: 28ms<br/>99%: 54ms   <br/>min: 1.256ms<br/>max: 79.409ms | requests: 367<br/>failures: 0<br/>req/s: 6.118req/s<br/>avg: 5.176ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 16ms<br/>98%: 29ms<br/>99%: 55ms   <br/>min: 1.353ms<br/>max: 73.315ms |
| e2e_match_mld | requests: 346<br/>failures: 0<br/>req/s: 5.769req/s<br/>avg: 5.235ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 13ms<br/>98%: 41ms<br/>99%: 54ms   <br/>min: 1.263ms<br/>max: 78.825ms | requests: 361<br/>failures: 0<br/>req/s: 6.024req/s<br/>avg: 5.188ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 15ms<br/>98%: 26ms<br/>99%: 54ms   <br/>min: 1.411ms<br/>max: 99.287ms |
| e2e_nearest_ch | requests: 452<br/>failures: 0<br/>req/s: 7.534req/s<br/>avg: 1.604ms<br/>50%: 2ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 2ms   <br/>min: 1.070ms<br/>max: 5.632ms | requests: 443<br/>failures: 0<br/>req/s: 7.385req/s<br/>avg: 1.609ms<br/>50%: 2ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 2ms   <br/>min: 1.044ms<br/>max: 2.439ms |
| e2e_nearest_mld | requests: 410<br/>failures: 0<br/>req/s: 6.836req/s<br/>avg: 1.891ms<br/>50%: 2ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 3ms   <br/>min: 1.000ms<br/>max: 53.821ms | requests: 431<br/>failures: 0<br/>req/s: 7.192req/s<br/>avg: 1.650ms<br/>50%: 2ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 3ms   <br/>min: 1.024ms<br/>max: 2.783ms |
| e2e_route_ch | requests: 423<br/>failures: 0<br/>req/s: 7.050req/s<br/>avg: 4.170ms<br/>50%: 4ms<br/>75%: 5ms<br/>95%: 6ms<br/>98%: 7ms<br/>99%: 9ms   <br/>min: 1.401ms<br/>max: 10.576ms | requests: 399<br/>failures: 0<br/>req/s: 6.651req/s<br/>avg: 4.549ms<br/>50%: 4ms<br/>75%: 5ms<br/>95%: 6ms<br/>98%: 7ms<br/>99%: 11ms   <br/>min: 1.654ms<br/>max: 122.872ms |
| e2e_route_mld | requests: 373<br/>failures: 0<br/>req/s: 6.219req/s<br/>avg: 5.008ms<br/>50%: 5ms<br/>75%: 5ms<br/>95%: 7ms<br/>98%: 10ms<br/>99%: 12ms   <br/>min: 1.619ms<br/>max: 96.637ms | requests: 394<br/>failures: 0<br/>req/s: 6.574req/s<br/>avg: 4.677ms<br/>50%: 4ms<br/>75%: 5ms<br/>95%: 6ms<br/>98%: 8ms<br/>99%: 10ms   <br/>min: 1.483ms<br/>max: 96.349ms |
| e2e_table_ch | requests: 387<br/>failures: 0<br/>req/s: 6.450req/s<br/>avg: 45.209ms<br/>50%: 27ms<br/>75%: 75ms<br/>95%: 130ms<br/>98%: 140ms<br/>99%: 140ms   <br/>min: 2.528ms<br/>max: 164.615ms | requests: 367<br/>failures: 0<br/>req/s: 6.118req/s<br/>avg: 49.607ms<br/>50%: 32ms<br/>75%: 83ms<br/>95%: 130ms<br/>98%: 140ms<br/>99%: 140ms   <br/>min: 2.580ms<br/>max: 161.204ms |
| e2e_table_mld | requests: 390<br/>failures: 0<br/>req/s: 6.503req/s<br/>avg: 75.482ms<br/>50%: 77ms<br/>75%: 110ms<br/>95%: 140ms<br/>98%: 150ms<br/>99%: 150ms   <br/>min: 4.564ms<br/>max: 161.799ms | requests: 381<br/>failures: 0<br/>req/s: 6.357req/s<br/>avg: 54.277ms<br/>50%: 35ms<br/>75%: 87ms<br/>95%: 140ms<br/>98%: 160ms<br/>99%: 180ms   <br/>min: 3.029ms<br/>max: 217.415ms |
| e2e_trip_ch | requests: 335<br/>failures: 0<br/>req/s: 5.584req/s<br/>avg: 16.960ms<br/>50%: 15ms<br/>75%: 23ms<br/>95%: 32ms<br/>98%: 33ms<br/>99%: 35ms   <br/>min: 1.565ms<br/>max: 41.705ms | requests: 353<br/>failures: 0<br/>req/s: 5.885req/s<br/>avg: 17.275ms<br/>50%: 16ms<br/>75%: 22ms<br/>95%: 32ms<br/>98%: 34ms<br/>99%: 35ms   <br/>min: 2.633ms<br/>max: 35.554ms |
| e2e_trip_mld | requests: 377<br/>failures: 0<br/>req/s: 6.286req/s<br/>avg: 21.645ms<br/>50%: 21ms<br/>75%: 27ms<br/>95%: 37ms<br/>98%: 38ms<br/>99%: 43ms   <br/>min: 3.859ms<br/>max: 52.583ms | requests: 347<br/>failures: 0<br/>req/s: 5.790req/s<br/>avg: 19.431ms<br/>50%: 17ms<br/>75%: 24ms<br/>95%: 35ms<br/>98%: 40ms<br/>99%: 52ms   <br/>min: 2.718ms<br/>max: 262.450ms |
| json-render | String: 6.83874ms<br/>Stringstream: 9.42975ms<br/>Vector: 6.99938ms | String: 6.61233ms<br/>Stringstream: 9.41045ms<br/>Vector: 6.927ms |
| match_ch | Default radius: <br/>4.39971ms/req at 82 coordinate<br/>0.053655ms/coordinate<br/>Radius 5m: <br/>4.38432ms/req at 82 coordinate<br/>0.0534673ms/coordinate<br/>Radius 10m: <br/>14.9984ms/req at 82 coordinate<br/>0.182907ms/coordinate<br/>Radius 15m: <br/>36.6858ms/req at 82 coordinate<br/>0.447387ms/coordinate<br/>Radius 30m: <br/>312.541ms/req at 82 coordinate<br/>3.81148ms/coordinate | Default radius: <br/>4.40002ms/req at 82 coordinate<br/>0.0536588ms/coordinate<br/>Radius 5m: <br/>4.3726ms/req at 82 coordinate<br/>0.0533244ms/coordinate<br/>Radius 10m: <br/>14.9772ms/req at 82 coordinate<br/>0.182649ms/coordinate<br/>Radius 15m: <br/>36.6045ms/req at 82 coordinate<br/>0.446397ms/coordinate<br/>Radius 30m: <br/>312.043ms/req at 82 coordinate<br/>3.80541ms/coordinate |
| match_mld | Default radius: <br/>2.79289ms/req at 82 coordinate<br/>0.0340597ms/coordinate<br/>Radius 5m: <br/>2.76467ms/req at 82 coordinate<br/>0.0337155ms/coordinate<br/>Radius 10m: <br/>10.1392ms/req at 82 coordinate<br/>0.123649ms/coordinate<br/>Radius 15m: <br/>25.7933ms/req at 82 coordinate<br/>0.314553ms/coordinate<br/>Radius 30m: <br/>302.272ms/req at 82 coordinate<br/>3.68624ms/coordinate | Default radius: <br/>2.89472ms/req at 82 coordinate<br/>0.0353015ms/coordinate<br/>Radius 5m: <br/>2.88337ms/req at 82 coordinate<br/>0.0351631ms/coordinate<br/>Radius 10m: <br/>10.8536ms/req at 82 coordinate<br/>0.132362ms/coordinate<br/>Radius 15m: <br/>28.039ms/req at 82 coordinate<br/>0.341939ms/coordinate<br/>Radius 30m: <br/>329.688ms/req at 82 coordinate<br/>4.02059ms/coordinate |
| osrm_contract | Time: 95.04s Peak RAM: 185.49MB | Time: 94.04s Peak RAM: 185.68MB |
| osrm_customize | Time: 1.32s Peak RAM: 115.02MB | Time: 1.34s Peak RAM: 115.02MB |
| osrm_extract | Time: 12.45s Peak RAM: 418.14MB | Time: 12.42s Peak RAM: 410.60MB |
| osrm_partition | Time: 2.29s Peak RAM: 148.46MB | Time: 2.25s Peak RAM: 150.45MB |
| packedvector | random write:<br/>std::vector 11841.7 ms<br/>util::packed_vector 77584.7 ms<br/>slowdown: 6.55183<br/>random read:<br/>std::vector 11502 ms<br/>util::packed_vector 32121.5 ms<br/>slowdown: 2.79269 | random write:<br/>std::vector 9834.68 ms<br/>util::packed_vector 82055.2 ms<br/>slowdown: 8.34345<br/>random read:<br/>std::vector 8536.84 ms<br/>util::packed_vector 33709.8 ms<br/>slowdown: 3.94874 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>514.283ms<br/>0.514283ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>353.754ms<br/>0.353754ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>628.087ms<br/>0.628087ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.559ms<br/>0.151559ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.3603ms<br/>0.0973603ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.979ms<br/>0.132979ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>150.479ms<br/>0.150479ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>96.9612ms<br/>0.0969612ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.259ms<br/>0.132259ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>511.879ms<br/>0.511879ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>352.403ms<br/>0.352403ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>630.176ms<br/>0.630176ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.984ms<br/>0.151984ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>96.809ms<br/>0.096809ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>131.784ms<br/>0.131784ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>148.425ms<br/>0.148425ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>96.6305ms<br/>0.0966305ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>130.961ms<br/>0.130961ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>638.348ms<br/>0.638348ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>435.959ms<br/>0.435959ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>809.74ms<br/>0.80974ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>264.405ms<br/>0.264405ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>161.915ms<br/>0.161915ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>287.528ms<br/>0.287528ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>264.671ms<br/>0.264671ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>159.96ms<br/>0.15996ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>283.35ms<br/>0.28335ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>650.198ms<br/>0.650198ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>442.618ms<br/>0.442618ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>827.181ms<br/>0.827181ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>284.729ms<br/>0.284729ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>162.978ms<br/>0.162978ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>303.87ms<br/>0.30387ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>283.513ms<br/>0.283513ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>162.839ms<br/>0.162839ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>304.112ms<br/>0.304112ms/req |
| rtree | 1 result:<br/>208.2ms ->  0.02082 ms/query<br/>10 results:<br/>242.303ms ->  0.0242303 ms/query | 1 result:<br/>207.261ms ->  0.0207261 ms/query<br/>10 results:<br/>242.239ms ->  0.0242239 ms/query |
</details>
<!-- BENCHMARK_RESULTS_END -->